+++ tests/modules/pantheon_advanced_page_cache_test/pantheon_advanced_page_cache_test.services.yml
@@ -1,0 +1,5 @@
+services:
+  pantheon_advanced_page_cache_test.cacheable_response_subscriber:
+    class: Drupal\pantheon_advanced_page_cache_test\EventSubscriber\CacheableResponseSubscriber
+    tags:
+      - { name: event_subscriber }
