--- tests/modules/pantheon_advanced_page_cache_tests/pantheon_advanced_page_cache_test.services.yml
@@ -1,6 +1,0 @@
-services:
-  pantheon_advanced_page_cache.cacheable_response_subscriber:
-    class: Drupal\pantheon_advanced_page_cache\EventSubscriber\CacheableResponseSubscriber
-    arguments: ['@logger.channel.pantheon_advanced_page_cache']
-    tags:
-      - { name: event_subscriber }
